qcom-legacy: msm8998: Sync with PDC config and cleanup
Another important distinction between k4.14 and k4.4 is that
wakegic is now configured as the interrupt parent for the SoC, so
it is important that calls needed to go to GIC gets passed
through.

One of these calls is set_wake, unfortunately, Sony's implementation
does not take into account that k4.4 mpm-of has a completely separate
algorithm and irq set that is being used for enable/disable calls for
wake. What happens as a result of this is that wake interrupts get
fully-enabled instead of getting listed as wake-enabled.

Let's use PDC's approach of bypassing set_wake to parent as it now
has been fixed in GIC. With the inversed hierarchical setup, it
may be fine now to no longer port the additional logic in k4.4.

In addition to this, cleanup other SoC quirks and also now setup
irq_enable/disable for MPM, matching PDC to better comply with
k4.14's config of having wakegic as parent IRQ for the SoC.

Signed-off-by: ederekun <sedrickvince@gmail.com>